requirements.txt
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dlib==19.24.0
easygui==0.98.3
face-recognition==1.3.0
face-recognition-models==0.3.0
findfile==1.7.10
fonttools==4.38.0
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
opencv-python-headless==4.6.0.66
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
six==1.16.0
termcolor==2.1.1
tk==0.1.0
